DHFPROD-1980: Fix gradle plugin tests

+++ marklogic-data-hub/src/main/java/com/marklogic/hub/impl/HubProjectImpl.java
@@ -417,7 +417,9 @@ public class HubProjectImpl implements HubProject {
                 }
             }
         }
-        Files.move(oldMappingsDir, newMappingsDirPath);
+        if (oldMappingsDir.toFile().exists()) {
+            Files.move(oldMappingsDir, newMappingsDirPath);
+        }
     }
 
     @Override  public String getHubModulesDeployTimestampFile() {
